Simplify RootExplore screen; show nearby data for species view
