+++ src/components/Explore/hooks/useMapLocation.js
@@ -20,6 +20,7 @@ const DELTA = 0.2;
 const useMapLocation = ( ): Object => {
   const { params } = useRoute( );
   const place = params?.place;
+  const worldwide = params?.worldwide;
   const realm = useRealm( );
   const { dispatch, state } = useExplore( );
   const [mapBoundaries, setMapBoundaries] = useState( null );
@@ -31,7 +32,7 @@ const useMapLocation = ( ): Object => {
     latitudeDelta: DELTA,
     longitudeDelta: DELTA
   } );
-  const [startAtNearby, setStartAtNearby] = useState( !state.swlat );
+  const [startAtNearby, setStartAtNearby] = useState( !state.swlat && !worldwide );
   const { t } = useTranslation( );
 
   const onPanDrag = ( ) => setShowMapBoundaryButton( true );
