+++ src/navigation/BottomTabNavigator/CustomTabBarContainer.js
@@ -8,7 +8,7 @@ import { useCurrentUser, useTranslation } from "sharedHooks";
 import CustomTabBar from "./CustomTabBar";
 
 const DRAWER_ID = "OPEN_DRAWER";
-const EXPLORE_SCREEN_ID = "Explore";
+const EXPLORE_SCREEN_ID = "RootExplore";
 const OBS_LIST_SCREEN_ID = "ObservationsStackNavigator";
 const NOTIFICATIONS_SCREEN_ID = "Notifications";
 
@@ -42,9 +42,7 @@ const CustomTabBarContainer = ( { navigation }: Props ): Node => {
       accessibilityHint: t( "Navigates-to-explore" ),
       size: 40,
       onPress: ( ) => {
-        navigation.navigate( "RootExplore", {
-          viewSpecies: true
-        } );
+        navigation.navigate( "RootExplore" );
         setActiveTab( EXPLORE_SCREEN_ID );
       },
       active: EXPLORE_SCREEN_ID === activeTab
